Move `TablesMigrate` to a separate module (#747)
